Refactor spec file name and check params class
